move ConfidenceInterval class
